requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
requests==2.31.0
python-whois==0.8.0
dnspython==2.4.2
validators==0.22.0
python-magic==0.4.27
pycryptodome==3.19.0
beautifulsoup4==4.12.2
Flask-Session==0.5.0
Flask-WTF==1.1.1
yara-python==4.3.1
scapy==2.5.0